requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
telethon==1.30.3
google-generativeai==0.3.2
supabase==2.3.4
reportlab==4.0.4
Pillow==10.0.1
matplotlib==3.10.3